+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.3.1 (unreleased)
 =====
 * Added missing dependency on `ocamlfind`
+* Start/stop log processing on `startup`/`cleanup`.
 
 0.3.0 (2022-03-04)
 =====